ABC and logger
I created an abstract base class to be used by all opponents types. The simulated mice and learners should be derived from the new class and implement the abstract methods. The __init__() function can be different for each type of opponent. I also made the location of the  experiment log file relative to the working directory and created an experiment specific file name